.ai-agent/requirements.txt
# AI博客生成系统依赖

# Claude API
anthropic>=0.40.0

# HTTP请求
requests>=2.31.0

# RSS解析
feedparser>=6.0.10

# YAML配置
PyYAML>=6.0.1

# 环境变量管理
python-dotenv>=1.0.0

# 邮件发送（系统自带）
# smtplib (标准库)
# email (标准库)

# 可选：代码格式化
black>=23.0.0
isort>=5.12.0

# 可选：定时任务
schedule>=1.2.0
